apps/backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Celery
redis==5.0.1
celery==5.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0

# Telegram
telegram==0.0.1
telebot==5.3.1
aiogram==3.4.1

# Utilities
python-dateutil==2.8.2
boto3==1.34.25
python-dotenv==1.0.0
email-validator==2.1.0.post1

# Validation
email-validator==2.1.0.post1

# CORS
starlette==0.35.1